refactor(api): move calculateDuplicateScore from praise transformers to praise/utils/score
