+++ packages/api/src/praise/utils/score.ts
@@ -1,5 +1,6 @@
 import { BadRequestError } from '@error/errors';
 import { settingValue } from '@shared/settings';
+import mongoose from 'mongoose';
 import { PraiseModel } from '../entities';
 import { PraiseDocument, Quantification } from '../types';
 import { getPraisePeriod } from './core';
@@ -59,4 +60,20 @@ export const calculatePraiseScore = async (
     praise.quantifications,
     duplicatePraisePercentage
   );
+};
+
+export const calculateDuplicateScore = async (
+  quantification: Quantification,
+  periodId: mongoose.Schema.Types.ObjectId
+): Promise<number> => {
+  const duplicatePraisePercentage = (await settingValue(
+    'PRAISE_QUANTIFY_DUPLICATE_PRAISE_PERCENTAGE',
+    periodId
+  )) as number;
+  if (!duplicatePraisePercentage)
+    throw new BadRequestError(
+      "Invalid setting 'PRAISE_QUANTIFY_DUPLICATE_PRAISE_PERCENTAGE'"
+    );
+
+  return Math.floor(quantification.score * duplicatePraisePercentage);
 };
